% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{dwei_series}
\alias{dwei_series}
\title{pdf for weibull series}
\usage{
dwei_series(t, shapes, scales)
}
\arguments{
\item{t}{series system lifetime}

\item{shapes}{shape parameters for weibull component lifetimes}

\item{scales}{scale parameters for weibull component lifetimes}
}
\description{
pdf for weibull series
}
