% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grad_ascent}
\alias{grad_ascent}
\title{Gradient ascent method}
\usage{
grad_ascent(
  fn,
  x0,
  gr,
  sup = function(x) TRUE,
  lr = 1,
  debug = FALSE,
  REPORT = 100L,
  r = 0.8,
  maxit = 100L
)
}
\arguments{
\item{fn}{function to be optimized}

\item{x0}{initial guess}

\item{gr}{gradient function}

\item{sup}{support function}

\item{lr}{learning rate}

\item{debug}{debug flag}

\item{REPORT}{report frequency}

\item{r}{line search decay}

\item{maxit}{maximum number of iterations}
}
\value{
list of converged, parameter, and number of iterations
}
\description{
Gradient ascent method
}
