% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_guo_weibull_series_data.R
\docType{data}
\name{guo_weibull_series_mle}
\alias{guo_weibull_series_mle}
\title{Weibull series MLE of \code{guo_weibull_series_md}}
\format{
A list with the following components:
\describe{
\item{mle}{The MLE of the shape and scale parameters}
\item{loglike}{The log-likelihood of the MLE}
\item{data}{The data used to compute the MLE}
\item{family}{The family of the likelihood model}
}
}
\usage{
guo_weibull_series_mle
}
\description{
When you use likelihood model that assumes a Weibull series system and
candidate sets represented by Boolean vectors (x1, x2, x3) that satisfy
conditions C1, C2, and C3, the MLE of the shape and
scale parameters are:
}
\details{
\itemize{
\item β1 = 1.2576
\item η1 = 994.3661
\item β2 = 1.1635
\item η2 = 908.9458
\item β3 = 1.1308
\item η3 = 840.1141
\item θ̂  = (β1, η1, β2, η2, β3, η3)
}

This has a log-likelihood of -228.6851.
}
\seealso{
\code{\link{guo_weibull_series_md}}
}
\keyword{datasets}
