% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_comp.R
\name{hazard_wei}
\alias{hazard_wei}
\title{Weibull Series System Hazard Function}
\usage{
hazard_wei(t, shape, scale)
}
\arguments{
\item{t}{time}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\description{
Weibull Series System Hazard Function
}
