% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_sim.R
\name{md_bernoulli_cand_c1_c2_c3}
\alias{md_bernoulli_cand_c1_c2_c3}
\title{Bernoulli candidate set model is a particular type of \emph{uninformed} model.
Note that we do not generate candidate sets with this function. See
\code{md_cand_sampler} for that.}
\usage{
md_bernoulli_cand_c1_c2_c3(df, p, control = list())
}
\arguments{
\item{df}{masked data.}

\item{p}{a vector of probabilities (p\link{j} is the probability that the jth
system will include a non-failed component in its candidate set,
assuming the jth system is not right-censored).}

\item{control}{\itemize{
\item \code{prob} - column prefix for component probabilities, defaults to
\code{q}, e.g., \verb{q1, q2, q3}.
\item \code{comp} - column prefix for component lifetimes, defaults to \code{t},
e.g., \verb{t1, t2, t3}.
\item \code{right_censoring_indicator} right-censoring indicator column name, if
TRUE, then the system lifetime is right-censored, otherwise it is
observed.
}}
}
\description{
This model satisfies conditions C1, C2, and C3.
The failed component will be in the corresponding candidate set with
probability 1, and the remaining components will be in the candidate set
with probability \code{p}.
}
