% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_masked_component_cause.R
\name{md_bernoulli_cand_c1_c2_c3_identifiable}
\alias{md_bernoulli_cand_c1_c2_c3_identifiable}
\title{Check if a masked data frame is identifiable.}
\usage{
md_bernoulli_cand_c1_c2_c3_identifiable(df, candset = "x")
}
\arguments{
\item{df}{masked data frame}

\item{candset}{column prefix for candidate sets, defaults to \code{x},
e.g., \verb{x1, x2, x3}.}
}
\value{
returns a string describing the data as yielding an identifiable
likelihood function, or one or more reasons why it is not identifiable.
}
\description{
We want to make sure that the right-censored, masked component cause of failure
data has enough information in it for the likelihood model to be uniquely
identified.
}
