% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_sim.R
\name{md_cand_sampler}
\alias{md_cand_sampler}
\title{Candidate set generator. Requires columns for component probabilities
e.g., \verb{q1,...,qm} where \code{qj} is the probability that the jth component
will be in the corresponding candidate set generated for that observation
in the \code{md} table.}
\usage{
md_cand_sampler(df, control = list())
}
\arguments{
\item{df}{(masked) data frame}

\item{control}{\itemize{
\item \code{prob} - column prefix for component probabilities, defaults to
\code{q}, e.g., \verb{q1, q2, q3}.
\item \code{candset} - column prefix for candidate sets (as Boolean matrix),
defaults to \code{x}, e.g., \verb{x1, x2, x3}.
}}
}
\description{
Candidate set generator. Requires columns for component probabilities
e.g., \verb{q1,...,qm} where \code{qj} is the probability that the jth component
will be in the corresponding candidate set generated for that observation
in the \code{md} table.
}
