% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_masked_component_cause.R
\name{md_decorate_component_cause_k}
\alias{md_decorate_component_cause_k}
\title{Add info about component cause of failure to a
masked data frame with latent components available.}
\usage{
md_decorate_component_cause_k(df, comp = "t")
}
\arguments{
\item{df}{masked data frame}

\item{comp}{column prefix for component lifetimes, defaults to \code{t},}
}
\value{
a data frame with a new \code{k} column
}
\description{
Add info about component cause of failure to a
masked data frame with latent components available.
}
