% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_masked_component_cause.R
\name{md_sample_candidates}
\alias{md_sample_candidates}
\title{Empirical sampler for C\link{i} | K\link{i} = k given a masked
data frame with component cause of failures shown.}
\usage{
md_sample_candidates(df, n, k, cause = "k", candset = "x")
}
\arguments{
\item{df}{masked data frame}

\item{cause}{column name for component cause of failure, defaults to \code{k}}

\item{candset}{column prefix for candidate sets, defaults to \code{x},
e.g., \verb{x1, x2, x3}.}
}
\value{
a data frame of candidate sets sampled from the empirical
distribution of C\link{i} | K\link{i} = k
}
\description{
In the likelihood model, we assume rather that
C\link{i} | T\link{i} = t, K\link{i} = k satisfies certain assumptions.
We're letting the data speak for itself, except for tractability
we're discarding the conditional relation on T\link{i} and only
keeping the conditional relation on K\link{i}, whatever it happens to be.
}
\details{
We could discretize the component lifetimes and then sample
from the empirical distribution of C\link{i} | T\link{i} = t, K\link{i} = k,
but that would be a lot of work and we'd have to choose a
discretization scheme. Instead, we'll just sample from the
empirical distribution of C\link{i} | K\link{i} = k.
}
