% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_md_wei_series_c1_c2_c3.R
\name{mle_cg_wei_series_md_c1_c2_c3}
\alias{mle_cg_wei_series_md_c1_c2_c3}
\title{optimization using conjugate gradient}
\usage{
mle_cg_wei_series_md_c1_c2_c3(
  df,
  theta0,
  reltol = 1e-10,
  parscale = NULL,
  maxit = 10000,
  ...,
  control = list()
)
}
\arguments{
\item{df}{data frame, right-censored lifetimes with masked component cause of failure}

\item{theta0}{initial parameter vector}

\item{reltol}{relative tolerance}

\item{parscale}{parameter scaling vector}

\item{maxit}{maximum number of iterations}

\item{...}{additional arguments passed to \code{optim}}

\item{control}{list of control parameters passed to \code{loglik_wei_series_md_c1_c2_c3}}
}
\value{
list with components:
\itemize{
\item \code{par} final parameter vector
\item \code{value} final negative log-likelihood
\item \code{counts} number of function evaluations, gradient evaluations, and Hessian evaluations
\item \code{convergence} convergence code
\item \code{message} convergence message
\item \code{hessian} estimated Hessian
}
}
\description{
optimization using conjugate gradient
}
