% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_mle_wei_series_c1_c2_c3.R
\name{mle_lbfgsb_wei_series_md_c1_c2_c3}
\alias{mle_lbfgsb_wei_series_md_c1_c2_c3}
\title{L-BFGS-B MLE solver for the Weibull series model with
masked component cause of failure with observation of exact system lifetime and
right censoring.}
\usage{
mle_lbfgsb_wei_series_md_c1_c2_c3(
  df,
  theta0,
  hessian = TRUE,
  ...,
  control = list()
)
}
\arguments{
\item{df}{data frame, right-censored lifetimes with masked component cause of failure}

\item{theta0}{initial parameter vector}

\item{...}{additional arguments passed to \code{loglik_wei_series_md_c1_c2_c3} and
\code{score_wei_series_md_c1_c2_c3}.}

\item{control}{list of control parameters passed to \code{optim} for \code{method == "L-BFGS-B"}}
}
\value{
list with components:
\itemize{
\item \code{par} final parameter vector
\item \code{value} final log-likelihood
\item \code{counts} number of function evaluations, gradient evaluations, and Hessian evaluations
\item \code{convergence} convergence code
\item \code{message} convergence message
\item \code{hessian} estimated Hessian
}
}
\description{
L-BFGS-B MLE solver for the Weibull series model with
masked component cause of failure with observation of exact system lifetime and
right censoring.
}
