% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{newton_raphson}
\alias{newton_raphson}
\title{Newton-Raphson method}
\usage{
newton_raphson(
  fn,
  x0,
  gr,
  hess,
  parscale = 1,
  tol = 1e-10,
  sup = function(x) TRUE,
  lr = 1,
  debug = FALSE,
  REPORT = 100L,
  r = 0.5,
  maxit = 1000L
)
}
\arguments{
\item{fn}{function to be optimized}

\item{x0}{initial guess}

\item{gr}{gradient function}

\item{hess}{hessian function}

\item{parscale}{parameter scaling}

\item{tol}{tolerance}

\item{sup}{support function}

\item{lr}{learning rate}

\item{debug}{debug flag}

\item{REPORT}{report frequency}

\item{r}{line search decay}

\item{maxit}{maximum number of iterations}
}
\value{
list of converged, parameter, number of iterations, hessian, and value
}
\description{
Newton-Raphson method
}
