% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{pwei_series}
\alias{pwei_series}
\title{The cumulative distribution function for Weibull series}
\usage{
pwei_series(t, scales, shapes)
}
\arguments{
\item{t}{series system lifetime}

\item{scales}{scale parameters for Weibull component lifetimes}

\item{shapes}{shape parameters for Weibull component lifetimes}
}
\description{
The cumulative distribution function for Weibull series
}
