% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_wei_series_c1_c2_c3.R
\name{score_wei_series_md_c1_c2_c3}
\alias{score_wei_series_md_c1_c2_c3}
\title{Generates a score function (gradient of the log-likelihood function) for a Weibull
series system with respect to parameter \code{theta} (shape, scale) for masked data
with candidate sets that satisfy conditions C1, C2, and C3 and right-censored
data.}
\usage{
score_wei_series_md_c1_c2_c3(df, theta, ..., control = list())
}
\arguments{
\item{df}{(masked) data frame}

\item{theta}{parameter vector (shape1, scale1, ..., shapem, scalem)}

\item{...}{additional arguments passed to \code{method.args} in \code{grad}}

\item{control}{list of control parameters. See \code{loglik_wei_series_md_c1_c2_c3}}
}
\value{
Score with respect to \code{theta} given \code{df}
}
\description{
Generates a score function (gradient of the log-likelihood function) for a Weibull
series system with respect to parameter \code{theta} (shape, scale) for masked data
with candidate sets that satisfy conditions C1, C2, and C3 and right-censored
data.
}
