% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{surv_wei_series}
\alias{surv_wei_series}
\title{Survival function for weibull series}
\usage{
surv_wei_series(t, scales, shapes)
}
\arguments{
\item{t}{series system lifetime}

\item{scales}{scale parameters for weibull component lifetimes}

\item{shapes}{shape parameters for weibull component lifetimes}
}
\description{
Survival function for weibull series
}
