% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series_comp.R
\name{wei_mttf}
\alias{wei_mttf}
\title{Mean-Time-To-Failure for Weibull}
\usage{
wei_mttf(shape, scale)
}
\arguments{
\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\description{
Mean-Time-To-Failure for Weibull
}
