% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{wei_series_cause}
\alias{wei_series_cause}
\title{The component cause of failure for the Weibull series distribution.}
\usage{
wei_series_cause(k, t = NA, shapes, scales)
}
\arguments{
\item{k}{component index}

\item{t}{lifetime of system, NA if unconditional distribution of K\link{i}}

\item{shapes}{shape parameters for Weibull component lifetimes}

\item{scales}{scale parameters for Weibull component lifetimes}
}
\value{
the probability that component k is the cause of failure,
for the unconditional random variable K\link{i}
}
\description{
The component cause of failure for the Weibull series distribution.
}
