% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{wei_series_mttf}
\alias{wei_series_mttf}
\title{The mean time to failure for the Weibull series system.}
\usage{
wei_series_mttf(shapes, scales)
}
\arguments{
\item{shapes}{shape parameters for Weibull component lifetimes}

\item{scales}{scale parameters for Weibull component lifetimes}
}
\description{
The mean time to failure for the Weibull series system.
}
