% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{Weibull series}
\alias{Weibull series}
\title{Weibull series system}
\description{
This file contains functions related to the Weibull series distribution.
Functions include simulation, pdf, cdf, quantile, and other related
functions for the Weibull series distribution.
}
\author{
Alex Towell
}
\keyword{distribution,}
\keyword{series,}
\keyword{statistics}
\keyword{weibull,}
