% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_guo_weibull_series_data.R
\docType{data}
\name{alex_weibull_series}
\alias{alex_weibull_series}
\title{This is the example series system used in the paper.
It has five components in series. It is based on the
example in Guo, table 2, but with two new components
added. Each component has approximately the same
reliability, so that there is no clear weakest link.}
\format{
A list with the following components:
\describe{
\item{shapes}{The shape parameters of the components}
\item{scales}{The scale parameters of the components}
\item{family}{The family of the likelihood model (weibull series)}
}
}
\usage{
alex_weibull_series
}
\description{
theta <- c(
k1 = 1.2576, lambda1 = 994.3661,
k2 = 1.1635, lambda2 = 908.9458,
k3 = 1.1308, lambda3 = 840.1141,
k4 = 1.1308, lambda4 = 940.1342,
k5 = 1.1308, lambda5 = 923.1631)
}
\seealso{
\code{\link{guo_weibull_series_md}}
}
\keyword{datasets}
