% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_bernoulli_masked_component_cause.R
\name{md_sample_candidates}
\alias{md_sample_candidates}
\title{Empirical sampler for C\link{i} | K\link{i} = k given a masked
data frame with component cause of failures shown.}
\usage{
md_sample_candidates(df, n, k = NA, cause = "k", candset = "x")
}
\arguments{
\item{df}{masked data frame}

\item{cause}{column name for component cause of failure, defaults to \code{k}}

\item{candset}{column prefix for candidate sets, defaults to \code{x},
e.g., \verb{x1, x2, x3}.}
}
\value{
a data frame of candidate sets sampled from the empirical
distribution of C\link{i} | K\link{i} = k
}
\description{
In the likelihood model, we assume rather that
C\link{i} | T\link{i} = t, K\link{i} = k satisfies certain assumptions.
For tractability and efficiency, we're discarding the conditional relation on
T\link{i} and only keeping the conditional relation on K\link{i} by using the
empirical distribution of C\link{i} | K\link{i} = k.
}
\details{
If k is not known, then we can sample from the empirical distribution of
C\link{i}, which may still be reasonable, since by Condition 2,
Pr{C\link{i} = c\link{i} | T\link{i} = t\link{i}, K\link{i} = j} =
Pr{C\link{i} = c\link{i} | T\link{i} = t\link{i}, K\link{i} = j'}  for all j, j' in c\link{i},
and so in general varying the component cause among the components
in the candidate set does not change the probability of the candidate set.

If the model is more informed, then dropping the conditional relation on
K\link{i} will lose information, and may bias the results in complex ways.
}
