% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_md_wei_series_c1_c2_c3.R
\name{mle_newton_wei_series_md_c1_c2_c3}
\alias{mle_newton_wei_series_md_c1_c2_c3}
\title{optimization using Newton-Raphson}
\usage{
mle_newton_wei_series_md_c1_c2_c3(
  df,
  theta0,
  lr = 1,
  tol = 1e-10,
  maxit = 10000,
  ...,
  control = list()
)
}
\arguments{
\item{df}{data frame, right-censored lifetimes with masked component cause of failure}

\item{theta0}{initial parameter vector}

\item{lr}{learning rate}

\item{tol}{tolerance}

\item{maxit}{maximum number of iterations}

\item{...}{additional arguments passed to \code{newton_raphson}}

\item{control}{list of control parameters passed to \code{loglik_wei_series_md_c1_c2_c3}}

\item{parscale}{parameter scaling}
}
\value{
list with components:
\itemize{
\item \code{par} final parameter vector
\item \code{value} final negative log-likelihood
\item \code{convergence} convergence code
\item \code{iter} number of iterations
\item \code{hessian} estimated Hessian
}
}
\description{
optimization using Newton-Raphson
}
