% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wei_series.R
\name{rwei_series}
\alias{rwei_series}
\title{Sampler for weibull series.}
\usage{
rwei_series(n, scales, shapes)
}
\arguments{
\item{n}{sample size}

\item{scales}{scale parameters for weibull component lifetimes}

\item{shapes}{shape parameters for weibull component lifetimes}
}
\description{
Sampler for weibull series.
}
