% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_wei_series_c1_c2_c3.R
\name{Weibull series MLE}
\alias{Weibull series MLE}
\title{Maximum likelihood estimation functions for Weibull series systems
from masked data.}
\description{
Functions include the log-likelihood, score, and MLE functions.
}
\details{
Masked component data approximately satisfies the following conditions:
C1: Pr(K in C) = 1
C2: Pr(C=c | K=j, T=t) = Pr(C=c | K=j', T=t)
for any j, j' in c.
C3: masking probabilities are independent of theta
}
\author{
Alex Towell
}
\keyword{data}
\keyword{distribution,}
\keyword{masked}
\keyword{series,}
\keyword{statistics,}
\keyword{weibull,}
